Feat/movie search

+++ app/src/main/java/com/mafaly/moviematch/db/daos/DuelDao.kt
@@ -10,16 +10,16 @@ import com.mafaly.moviematch.db.entities.DuelEntity
 @Dao
 interface DuelDao {
     @Query("SELECT * FROM duel_entity WHERE duel_game_id = :gameId")
-    fun getDuelsForGame(gameId: Int): List<DuelEntity>
+    fun getDuelsForGame(gameId: Long): List<DuelEntity>
 
     @Query("SELECT * FROM duel_entity WHERE duel_game_id = :gameId and duel_movie_1_id != null and duel_movie_2_id == null")
-    fun getIncompleteDuelsForGame(gameId: Int): List<DuelEntity>
+    fun getIncompleteDuelsForGame(gameId: Long): List<DuelEntity>
 
     @Query("SELECT * FROM duel_entity WHERE duel_game_id = :gameId and duel_winner_id == null")
-    fun getUnfinishedDuelsForGame(gameId: Int): List<DuelEntity>
+    fun getUnfinishedDuelsForGame(gameId: Long): List<DuelEntity>
 
     @Query("SELECT * FROM duel_entity WHERE id = :duelId")
-    fun getDuel(duelId: Int): DuelEntity
+    fun getDuel(duelId: Long): DuelEntity
 
     @Update
     fun updateDuel(duel: DuelEntity)
@@ -28,5 +28,5 @@ interface DuelDao {
     fun insertDuel(duel: DuelEntity)
 
     @Query("DELETE FROM duel_entity WHERE duel_game_id = :gameId")
-    fun deleteDuelsForGame(gameId: Int)
+    fun deleteDuelsForGame(gameId: Long)
 }
